fix: #5 북리뷰 패키지 구조 변경
