+++ backend/src/main/java/com/cbu/backend/book/controller/BookController.java
@@ -1,0 +1,7 @@
+package com.cbu.backend.book.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class BookController {
+}
