+++ backend/src/main/java/com/cbu/backend/book/dto/request/CreateBookRequestDTO.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.book.dto.request;
+
+public class CreateBookRequestDTO {
+}
