+++ backend/src/main/java/com/cbu/backend/book/dto/response/BookResponseDTO.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.book.dto.response;
+
+public class BookResponseDTO {
+}
