+++ backend/src/main/java/com/cbu/backend/book/entity/Book.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.book.entity;
+
+public class Book {
+}
