+++ backend/src/main/java/com/cbu/backend/book/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package com.cbu.backend.book.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BookRepository {
+}
