+++ backend/src/main/java/com/cbu/backend/book/service/BookService.java
@@ -1,0 +1,7 @@
+package com.cbu.backend.book.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BookService {
+}
