+++ backend/src/main/java/com/cbu/backend/bookReview/controller/BookReviewController.java
@@ -1,0 +1,7 @@
+package com.cbu.backend.bookReview.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class BookReviewController {
+}
