+++ backend/src/main/java/com/cbu/backend/bookReview/dto/request/CreateBookReviewRequestDTO.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.bookReview.dto.request;
+
+public class CreateBookReviewRequestDTO {
+}
