+++ backend/src/main/java/com/cbu/backend/bookReview/dto/response/BookReviewResponseDTO.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.bookReview.dto.response;
+
+public class BookReviewResponseDTO {
+}
