+++ backend/src/main/java/com/cbu/backend/bookReview/entity/BookReview.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.bookReview.entity;
+
+public class BookReview {
+}
