+++ backend/src/main/java/com/cbu/backend/bookReview/repository/BookReviewRepository.java
@@ -1,0 +1,7 @@
+package com.cbu.backend.bookReview.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BookReviewRepository {
+}
