+++ backend/src/main/java/com/cbu/backend/bookReview/service/BookReviewService.java
@@ -1,0 +1,7 @@
+package com.cbu.backend.bookReview.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BookReviewService {
+}
