+++ backend/src/main/java/com/cbu/backend/comment/dto/request/CreateCommentRequestDTO.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.comment.dto.request;
+
+public class CreateCommentRequestDTO {
+}
