+++ backend/src/main/java/com/cbu/backend/comment/dto/response/CommentResponseDTO.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.comment.dto.response;
+
+public class CommentResponseDTO {
+}
