+++ backend/src/main/java/com/cbu/backend/comment/entity/Comment.java
@@ -1,0 +1,4 @@
+package com.cbu.backend.comment.entity;
+
+public class Comment {
+}
