+++ backend/src/main/java/com/cbu/backend/comment/repository/CommentRepository.java
@@ -1,0 +1,7 @@
+package com.cbu.backend.comment.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CommentRepository {
+}
