+++ backend/src/main/java/com/cbu/backend/comment/service/CommentService.java
@@ -1,0 +1,7 @@
+package com.cbu.backend.comment.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CommentService {
+}
