--- backend/src/main/java/com/cbu/backend/controller/BookController.java
@@ -1,7 +1,0 @@
-package com.cbu.backend.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class BookController {
-}
