--- backend/src/main/java/com/cbu/backend/dto/book/request/CreateBookRequestDTO.java
@@ -1,4 +1,0 @@
-package com.cbu.backend.dto.book.request;
-
-public class CreateBookRequestDTO {
-}
