--- backend/src/main/java/com/cbu/backend/dto/book/response/BookResponseDTO.java
@@ -1,4 +1,0 @@
-package com.cbu.backend.dto.book.response;
-
-public class BookResponseDTO {
-}
