--- backend/src/main/java/com/cbu/backend/dto/bookReview/request/CreateBookReviewRequestDTO.java
@@ -1,4 +1,0 @@
-package com.cbu.backend.dto.bookReview.request;
-
-public class CreateBookReviewRequestDTO {
-}
