--- backend/src/main/java/com/cbu/backend/dto/bookReview/response/BookReviewResponseDTO.java
@@ -1,4 +1,0 @@
-package com.cbu.backend.dto.bookReview.response;
-
-public class BookReviewResponseDTO {
-}
