--- backend/src/main/java/com/cbu/backend/dto/comment/request/CreateCommentRequestDTO.java
@@ -1,4 +1,0 @@
-package com.cbu.backend.dto.comment.request;
-
-public class CreateCommentRequestDTO {
-}
