--- backend/src/main/java/com/cbu/backend/dto/comment/response/CommentResponseDTO.java
@@ -1,4 +1,0 @@
-package com.cbu.backend.dto.comment.response;
-
-public class CommentResponseDTO {
-}
