--- backend/src/main/java/com/cbu/backend/entity/Book.java
@@ -1,4 +1,0 @@
-package com.cbu.backend.entity;
-
-public class Book {
-}
