--- backend/src/main/java/com/cbu/backend/repository/CommentRepository.java
@@ -1,7 +1,0 @@
-package com.cbu.backend.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CommentRepository {
-}
