--- backend/src/main/java/com/cbu/backend/service/BookReviewService.java
@@ -1,7 +1,0 @@
-package com.cbu.backend.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class BookReviewService {
-}
